+++ src/createWebpackConfig.js
@@ -421,6 +421,11 @@ export const COMPAT_CONFIGS = {
       'react/lib/ReactContext': true
     }
   },
+  'json-schema': {
+    module: {
+      noParse: [/node_modules[/\\]json-schema[/\\]lib[/\\]validate\.js/]
+    }
+  },
   moment({locales}) {
     if (!Array.isArray(locales)) {
       console.error(red("nwb: webpack.compat.moment config must provide a 'locales' Array"))
@@ -429,7 +434,7 @@ export const COMPAT_CONFIGS = {
     return {
       plugins: [
         new webpack.ContextReplacementPlugin(
-          /moment[\\\/]locale$/,
+          /moment[/\\]locale$/,
           new RegExp(`^\\.\\/(${locales.join('|')})$`)
         )
       ]
@@ -437,7 +442,7 @@ export const COMPAT_CONFIGS = {
   },
   sinon: {
     module: {
-      noParse: /[/\\]sinon\.js/
+      noParse: [/[/\\]sinon\.js/]
     },
     resolve: {
       alias: {